+++ src/main/java/de/codescape/jira/plugins/scrumpoker/condition/ScrumPokerForIssueCondition.java
@@ -3,6 +3,7 @@ package de.codescape.jira.plugins.scrumpoker.condition;
 import com.atlassian.jira.issue.CustomFieldManager;
 import com.atlassian.jira.issue.Issue;
 import com.atlassian.jira.issue.fields.CustomField;
+import com.atlassian.jira.plugin.webfragment.conditions.AbstractIssueWebCondition;
 import com.atlassian.jira.plugin.webfragment.conditions.AbstractWebCondition;
 import com.atlassian.jira.plugin.webfragment.model.JiraHelper;
 import com.atlassian.jira.user.ApplicationUser;
@@ -12,7 +13,7 @@ import de.codescape.jira.plugins.scrumpoker.persistence.StoryPointFieldSupport;
  * This condition is used to decide whether a button to start a Scrum poker session should be displayed or not for the
  * given issue. The issue must be editable and have the custom field that is configured for Scrum poker estimations.
  */
-public class ScrumPokerForIssueCondition extends AbstractWebCondition {
+public class ScrumPokerForIssueCondition extends AbstractIssueWebCondition {
 
     private final CustomFieldManager customFieldManager;
     private final StoryPointFieldSupport storyPointFieldSupport;
@@ -24,9 +25,8 @@ public class ScrumPokerForIssueCondition extends AbstractWebCondition {
     }
 
     @Override
-    public boolean shouldDisplay(ApplicationUser applicationUser, JiraHelper jiraHelper) {
-        Issue currentIssue = (Issue) jiraHelper.getContextParams().get("issue");
-        return currentIssue.isEditable() && hasStoryPointField(currentIssue);
+    public boolean shouldDisplay(ApplicationUser applicationUser, Issue issue, JiraHelper jiraHelper) {
+        return issue.isEditable() && hasStoryPointField(issue);
     }
 
     private boolean hasStoryPointField(Issue currentIssue) {